fix(side_shift): split parameters definition
