+++ planning/behavior_path_planner/include/behavior_path_planner/scene_module/side_shift/side_shift_module.hpp
@@ -17,6 +17,7 @@
 
 #include "behavior_path_planner/scene_module/scene_module_interface.hpp"
 #include "behavior_path_planner/util/path_shifter/path_shifter.hpp"
+#include "behavior_path_planner/util/side_shift/side_shift_parameters.hpp"
 
 #include <rclcpp/rclcpp.hpp>
 
@@ -34,25 +35,6 @@ using autoware_auto_planning_msgs::msg::PathWithLaneId;
 using geometry_msgs::msg::Pose;
 using nav_msgs::msg::OccupancyGrid;
 using tier4_planning_msgs::msg::LateralOffset;
-
-enum class SideShiftStatus { STOP = 0, BEFORE_SHIFT, SHIFTING, AFTER_SHIFT };
-
-struct SideShiftParameters
-{
-  double time_to_start_shifting;
-  double min_distance_to_start_shifting;
-  double shifting_lateral_jerk;
-  double min_shifting_distance;
-  double min_shifting_speed;
-  double drivable_area_resolution;
-  double drivable_area_width;
-  double drivable_area_height;
-  double shift_request_time_limit;
-  // drivable area expansion
-  double drivable_area_right_bound_offset;
-  double drivable_area_left_bound_offset;
-  std::vector<std::string> drivable_area_types_to_skip;
-};
 
 class SideShiftModule : public SceneModuleInterface
 {
